refactor: rename user model
